#181 use /api/v1/auth/verify to probe whether WS will be successful
The reason is that WS gives you no information on why establishing
the connection failed. But we need to react differently on a 403 reason.
If we get 403 then we need to show the login form and allow the user to
login. If the probe call returns 200 then we attempt to connect the WS.

We also reworked the authentication related flags to have the following
two: authRequired and authEnabled. Both are options. authEnabled is
set through the aboutInfo coming from the WS, authRequired is determined
through the response (e.g. 403) from the verify response.

Current problems still
- After logging in you need to wait up to 5 seconds for the next probe
- When disconnected, the model is not reset and you see no difference
  in the UI