Parse and inject nomad configuration
Add a NomadModule which loads the configuration and provides it for
dependency injection.  Change services to receive a nomad configuration
instead of a play configuration via dependency injection.

Remove constants for config key and config defaults.  The config key is
only use once now, so the constant becomes redundant.

The default moves to "reference.conf" where typesafe config
automatically merges it with "application.conf", in line with typesafe
config best practices.

Eventually this should be the way to handle all configuration.